requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
affine=2.2.1=py36_0
appdirs=1.4.3=py36h28b3542_0
asn1crypto=0.24.0=py36_0
atomicwrites=1.1.5=py36_0
attrs=18.1.0=py36_0
automat=0.7.0=py36_0
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=2.1.3=py36_0
boto3=1.7.62=py36_1
botocore=1.10.62=py36_0
bzip2=1.0.6=h14c3975_5
ca-certificates=2018.03.07=0
cairo=1.14.12=h8948797_3
certifi=2018.4.16=py36_0
cffi=1.11.5=py36h9745a5d_0
click=6.7=py36_0
click-plugins=1.0.3=py36_0
cligj=0.4.0=py36_0
constantly=15.1.0=py36h28b3542_0
cryptography=2.2.2=py36h14c3975_0
curl=7.61.0=h84994c4_0
cycler=0.10.0=py36_0
dbus=1.13.2=h714fa37_1
decorator=4.3.0=py36_0
descartes=1.1.0=py36_0
docutils=0.14=py36_0
entrypoints=0.2.3=py36_2
expat=2.2.5=he0dffb1_0
fiona=1.7.12=py36h3f37509_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_0
freexl=1.0.5=h14c3975_0
gdal=2.2.2=py36hc209d97_1
geopandas=0.3.0=py36_0
geos=3.6.2=heeff764_2
giflib=5.1.4=h14c3975_1
glib=2.56.1=h000015b_0
gmp=6.1.2=h6c8ec71_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
hdf4=4.2.13=h3ca952b_2
hdf5=1.10.2=hba1933b_1
html5lib=1.0.1=py36_0
hyperlink=18.0.0=py36_0
icu=58.2=h9c2bf20_1
idna=2.7=py36_0
incremental=17.5.0=py36_0
intel-openmp=2018.0.3=0
ipykernel=4.8.2=py36_0
ipython=6.5.0=py36_0
ipython_genutils=0.2.0=py36_0
jedi=0.12.1=py36_0
jinja2=2.10=py36_0
jmespath=0.9.3=py36_0
jpeg=9b=h024ee3a_2
json-c=0.13.1=h1bed415_0
jsonschema=2.6.0=py36_0
jupyter_client=5.2.3=py36_0
jupyter_core=4.4.0=py36_0
jupyterlab=0.33.4=py36_0
jupyterlab_launcher=0.11.2=py36_0
kealib=1.4.7=h77bc034_6
kiwisolver=1.0.1=py36hf484d3e_0
krb5=1.16.1=hc83ff2d_6
libboost=1.67.0=h46d08c1_4
libcurl=7.61.0=h1ad7b7a_0
libdap4=3.19.1=h6ec2957_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.2.0=hdf63c60_3
libgdal=2.2.4=h6f639c0_1
libgfortran-ng=7.2.0=hdf63c60_3
libkml=1.3.0=h590aaf7_4
libnetcdf=4.6.1=h13459d8_0
libpng=1.6.34=hb9fc6fc_0
libpq=10.4=h1ad7b7a_0
libsodium=1.0.16=h1bed415_0
libspatialindex=1.8.5=h20b78c2_2
libspatialite=4.3.0a=he475c7f_19
libssh2=1.8.0=h9cfc8f7_4
libstdcxx-ng=7.2.0=hdf63c60_3
libtiff=4.0.9=he85c1e1_1
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
markupsafe=1.0=py36h14c3975_1
matplotlib=2.2.2=py36hb69df0a_2
mistune=0.8.3=py36h14c3975_1
mkl=2018.0.3=1
mkl_fft=1.0.4=py36h4414c95_1
mkl_random=1.0.1=py36h4414c95_1
more-itertools=4.2.0=py36_0
munch=2.3.2=py36_0
nbconvert=5.3.1=py36_0
nbformat=4.4.0=py36_0
ncurses=6.1=hf484d3e_0
notebook=5.6.0=py36_0
numpy=1.15.0=py36h1b885b7_0
numpy-base=1.15.0=py36h3dfced4_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.0.2o=h14c3975_1
pandas=0.23.3=py36h04863e7_0
pandoc=2.2.1=h629c226_0
pandocfilters=1.4.2=py36_1
parso=0.3.1=py36_0
pcre=8.42=h439df22_0
pexpect=4.6.0=py36_0
pickleshare=0.7.4=py36_0
pip=10.0.1=py36_0
pixman=0.34.0=hceecf20_3
pluggy=0.6.0=py36_0
poppler=0.65.0=h581218d_1
poppler-data=0.4.9=0
progressbar2=3.37.1=py36_0
proj4=5.0.1=h14c3975_0
prometheus_client=0.3.1=py36_0
prompt_toolkit=1.0.15=py36_0
psycopg2=2.7.5=py36hb7f436b_0
ptyprocess=0.6.0=py36_0
py=1.5.4=py36_0
pyasn1=0.4.4=py36_0
pyasn1-modules=0.2.2=py36_0
pycparser=2.18=py36_1
pygments=2.2.0=py36_0
pyopenssl=18.0.0=py36_0
pyparsing=2.2.0=py36_1
pyproj=1.9.5.1=py36h7b21b82_1
pyqt=5.9.2=py36h22d08a2_0
pysal=1.14.4.post1=py36_1
pytest=3.6.4=py36_0
pytest-runner=4.2=py36_0
python=3.6.4=hc3d631a_3
python-dateutil=2.7.3=py36_0
python-utils=2.3.0=py36_0
pytz=2018.5=py36_0
pyzmq=17.0.0=py36h14c3975_3
qt=5.9.6=h52aff34_0
rasterio=0.36.0=py36h3f37509_2
readline=7.0=ha6073c6_4
rtree=0.8.3=py36_0
s3transfer=0.1.13=py36_0
scipy=1.1.0=py36hc49cb51_0
send2trash=1.5.0=py36_0
service_identity=17.0.0=py36h28b3542_0
setuptools=39.2.0=py36_0
shapely=1.6.4=py36h7ef4460_0
simplegeneric=0.8.1=py36_2
sip=4.19.8=py36hf484d3e_0
six=1.11.0=py36_1
snuggs=1.4.2=py36_0
sqlalchemy=1.2.10=py36h14c3975_0
sqlite=3.24.0=h84994c4_0
terminado=0.8.1=py36_1
testpath=0.3.1=py36_0
tk=8.6.7=hc745277_3
tornado=5.0.2=py36h14c3975_0
traitlets=4.3.2=py36_0
twisted=18.4.0=py36h14c3975_0
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_1
wheel=0.31.1=py36_0
xerces-c=3.2.1=hac72e42_0
xz=5.2.4=h14c3975_4
zeromq=4.2.5=hf484d3e_0
zlib=1.2.11=ha838bed_2
zope=1.0=py36_0
zope.interface=4.5.0=py36h14c3975_0
